chore(edit): 메타데이터 생성, EditorContainer 추가
