+++ app/edit/[title]/page.tsx
@@ -1,5 +1,9 @@
-import Editor from "@/components/Editor";
-import { useDocsByTitleQuery } from "@/services/docs/docs.query";
+import getQueryClient from "@/app/getQueryClient";
+import EditorContainer from "@/components/Editor";
+import { docsQuery } from "@/services/docs/docs.query";
+import { generateOpenGraph } from "@/utils";
+import { HydrationBoundary, dehydrate } from "@tanstack/react-query";
+import { Metadata } from "next";
 import React from "react";
 
 interface PageProps {
@@ -8,12 +12,24 @@ interface PageProps {
   };
 }
 
+export const generateMetadata = async ({ params: { title } }: PageProps): Promise<Metadata> => {
+  const queryClient = getQueryClient();
+  const data = await queryClient.fetchQuery(docsQuery.title(title));
+
+  return generateOpenGraph({
+    title: `문서 편집#${data.title}`,
+    description: `${data.title} 문서 편집 페이지입니다.`,
+  });
+};
+
 const Page = async ({ params: { title } }: PageProps) => {
-  const docs = await useDocsByTitleQuery({ title });
+  const queryClient = getQueryClient();
+  await queryClient.prefetchQuery(docsQuery.title(title));
+
   return (
-    <div>
-      <Editor {...docs} mode="EDIT" />
-    </div>
+    <HydrationBoundary state={dehydrate(queryClient)}>
+      <EditorContainer title={title} mode="EDIT" />
+    </HydrationBoundary>
   );
 };
